requirements.txt
torch
torchvision
matplotlib
streamlit
streamlit-drawable-canvas
opencv-python
Pillow
